requirements.txt
Django==5.2.8
pillow==12.0.0
django-tinymce==5.0.0